CORE-514
manually adding in the HDB-Helium module as it is in a private repo and for broad release people will not have credentials to the repo.  this may change but for now we need to manually include the node file & module.  the repo for the module can be found here: https://github.com/HarperDB/harperdb_helium/tree/levyx_branch