CORE-991
changed jobs to spawn a new process which will run the intended job without impacting a main HDB process. effectively running the job as a true background process.

broke out the spawnSCConnection function from hdb_express into it's own module so it can be reused for jobThread.js.  the reason jobThread.jsneeds it's own connection is to sync csv loads to the cluster.