## CORE-1022 - Bug fix where a value like "1e10", "89e15636", "22e+10", etc... is interpreted as a number incorrectly. ##
Updated autocast to exclude a value that is detected to be a number that has a "e" in the value, this is to eliminate values that are interpteted to be exponents.
Added unit tests to test this scenario
Updated SQLSearch to not autocast the values stated above into a NumValue.